Added ahci read bare bones

+++ kernel/include/driver/ahci/ahci.h
@@ -21,7 +21,7 @@ void reset_hba();
 
 uint8_t send_ahci_command(uint8_t port, Register_H2D_FIS* command_fis, uint16_t prd_count, HBA_prdt_item* prdtp, uint16_t flags);
 
-void ahci_read(uint8_t port, uint64_t start_lba, uint32_t count, uint16_t* data);
+void ahci_read(uint8_t port, uint64_t start_lba, uint16_t count, uint16_t* data);
 
 void init_hba_port(uint64_t port);
 void init_ahci();
